farm_app/requirements-1.txt
greenlet==3.1.1
RPi.GPIO==0.7.1
sensor==6
smbus2==0.5.0
spidev==3.6
SQLAlchemy==2.0.37
typing_extensions==4.12.2
